fix: add ordering and filtering capabilities in TagService and introduce QueryTag class
