+++ src/Services/TagService.php
@@ -4,6 +4,7 @@ namespace CSlant\Blog\Api\Services;
 
 use Botble\Base\Models\BaseQueryBuilder;
 use Botble\Language\Facades\Language;
+use CSlant\Blog\Api\Supports\Queries\QueryTag;
 use CSlant\Blog\Core\Http\Responses\Base\BaseHttpResponse;
 use CSlant\Blog\Core\Models\Tag;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
@@ -29,15 +30,11 @@ class TagService
      */
     public function getFilters(array $filters): LengthAwarePaginator
     {
-        $data = Tag::query()
-            ->withCount('posts');
+        $query = Tag::query()->withCount('posts')->with([]);
 
-        if ($filters['search'] !== null) {
-            $keyword = isset($filters['search']) ? (string) $filters['search'] : null;
-            $data = $this->search($data, $keyword);
-        }
+        $query = QueryTag::setBaseCustomFilterQuery($query, $filters);
 
-        $data = $data
+        $data = $query
             ->wherePublished()
             ->orderBy(
                 Arr::get($filters, 'order_by', 'posts_count'),
@@ -45,33 +42,5 @@ class TagService
             );
 
         return $data->paginate((int) $filters['per_page']);
-    }
-
-    /**
-     * @param  BaseQueryBuilder|Builder<Model>  $model
-     * @param  null|string  $keyword
-     *
-     * @return BaseQueryBuilder|Builder<Model>
-     */
-    protected function search(Builder|BaseQueryBuilder $model, ?string $keyword): Builder|BaseQueryBuilder
-    {
-        if (!$model instanceof BaseQueryBuilder || !$keyword) {
-            return $model;
-        }
-
-        if (is_plugin_active('language')
-            && is_plugin_active('language-advanced')
-            && Language::getCurrentLocale() != Language::getDefaultLocale()
-        ) {
-            return $model
-                ->whereHas('translations', function (BaseQueryBuilder $query) use ($keyword): void {
-                    $query->addSearch('name', $keyword, false, false);
-                });
-        }
-
-        return $model
-            ->where(function (BaseQueryBuilder $query) use ($keyword): void {
-                $query->addSearch('name', $keyword, false, false);
-            });
     }
 }
